Refactor tool executor to use secure registry for built-in tools
- Extract tool implementations to `lib/toolImplementations.ts`
- Update `lib/toolExecutor.ts` to execute built-in tools directly, bypassing `new Function`
- Add `isCustom` flag to `ToolDefinition` to identify user-modified tools
- Maintain `new Function` fallback only for tools explicitly marked as custom
- Update `ToolEditor` to flag tools as custom on save
- Fixes Critical RCE vulnerability for trusted code paths
